Fix crash while compiling game (SCPCB)

+++ gxruntime/gxscene.cpp
@@ -242,17 +242,17 @@ void gxScene::setZMode(){
 void gxScene::setLights(){
 	if( fx & FX_FULLBRIGHT ){
 		//no lights on
-		for( int n=0;n<_curLights.size();++n ) dir3dDev->LightEnable( n,false );
+		for( int n=0;n<(int)_curLights.size();++n ) dir3dDev->LightEnable( n,false );
 	}else if( fx & FX_CONDLIGHT ){
 		//some lights on
-		for( int n=0;n<_curLights.size();++n ){
+		for( int n=0;n<(int)_curLights.size();++n ){
 			gxLight *light=_curLights[n];
 			bool enable=light->d3d_light.dltType!=D3DLIGHT_DIRECTIONAL;
 			dir3dDev->LightEnable( n,enable );
 		}
 	}else{
 		//all lights on
-		for( int n=0;n<_curLights.size();++n ) dir3dDev->LightEnable( n,true );
+		for( int n=0;n<(int)_curLights.size();++n ) dir3dDev->LightEnable( n,true );
 	}
 }
 
@@ -420,7 +420,7 @@ void gxScene::setRenderState( const RenderState &rs ){
 	if( rs.alpha!=material.diffuse.a ){
 		material.diffuse.a=rs.alpha;
 		if( rs.fx&FX_ALPHATEST ){
-			int alpharef=(rs.fx&FX_VERTEXALPHA)?0:128*rs.alpha;
+			int alpharef=(int)((rs.fx&FX_VERTEXALPHA)?0:128*rs.alpha);
 			setRS( D3DRENDERSTATE_ALPHAREF,alpharef );
 		}
 		setmat=true;
@@ -484,7 +484,7 @@ void gxScene::setRenderState( const RenderState &rs ){
 		}
 		if( t&FX_ALPHATEST ){
 			if( fx&FX_ALPHATEST ){
-				int alpharef=(rs.fx&FX_VERTEXALPHA)?0:128*rs.alpha;
+				int alpharef=(int)((rs.fx&FX_VERTEXALPHA)?0:128*rs.alpha);
 				setRS( D3DRENDERSTATE_ALPHAREF,alpharef );
 			}
 			setRS( D3DRENDERSTATE_ALPHATESTENABLE,fx & FX_ALPHATEST ? true : false );
@@ -552,7 +552,7 @@ bool gxScene::begin( const vector<gxLight*> &lights ){
 	//set light states
 	_curLights.clear();
 	for( n=0;n<8;++n ){
-		if( n<lights.size() ){
+		if( n<(int)lights.size() ){
 			_curLights.push_back( lights[n] );
 			dir3dDev->SetLight( n,&_curLights[n]->d3d_light );
 		}else{
